Extract JoinTestUtils
